added Apartment page and Bradcumb
