+++ src/App.js
@@ -4,6 +4,7 @@ import Home from './Components/Home/Home/Home';
 import NotFound from './Components/NotFound/NotFound';
 import Login from './Components/Login/Login';
 import { createContext, useState } from 'react';
+import Apartment from './Components/Apartment/Apartment/Apartment';
 
 export const UserContext = createContext();
 function App() {
@@ -18,8 +19,8 @@ function App() {
 					<Route exact path='/'>
 						<Home />
 					</Route>
-					<Route path='/apartment:id'>
-						<NotFound />
+					<Route path='/apartment/:id'>
+						<Apartment />
 					</Route>
 					<Route path='/login'>
 						<Login />
