+++ src/Components/Apartment/Apartment/Apartment.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../../Home/Header/Header';
+import Bradcumb from '../Bradcumb/Breadcumb';
+
+const Apartment = () => {
+	return (
+		<div>
+			<Header />
+			<Bradcumb />
+		</div>
+	);
+};
+
+export default Apartment;
