+++ src/Components/Apartment/Bradcumb/Breadcumb.scss
@@ -8,7 +8,7 @@
         right: 0px;
         bottom: 0px;
         left: 0px;
-        background-image:linear-gradient(rgba(0,0,0,0.6),rgba(0,0,0,0.5)), url(../../image/images/image1.png);
+        background-image:linear-gradient(rgba(0,0,0,0.6),rgba(0,0,0,0.5)), url(../../../images/apartmet-images/Rectangle-13.jpg);
         background-size: cover;
         background-position-y: 73%;
         background-repeat: no-repeat;
