--- src/Components/HeroArea/HeroArea.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import SearchBox from '../SearchBox/SearchBox';
-import './Heroarea.scss';
-
-const HeroArea = () => {
-    return (
-        <div>
-            <div className="HeroAreaStyle">
-                <SearchBox/>
-            </div>
-        </div>
-    );
-};
-
-export default HeroArea;
