+++ src/Components/Home/HeroArea/Heroarea.scss
@@ -1,0 +1,21 @@
+.HeroAreaStyle {
+    height: 400px;
+    position: relative;
+    z-index: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+
+    &::before {
+        content: '';
+        position: absolute;
+        top: 0px;
+        right: 0px;
+        height: 100%;
+        width: 100%;
+        background-color: #000;;
+        z-index: -1;
+        opacity: .75;
+    }
+}
