+++ src/Components/Home/SearchBox/SearchBox.scss
@@ -1,0 +1,12 @@
+
+.SearchBoxStyle {
+    color: white;
+}
+.formStyle{
+    width: 450px;
+    height: 45px;
+    border-radius: 5px;
+    padding-left: 15px;
+    border:none;
+    margin-right: 15px !important;
+}
