--- src/Components/SearchBox/SearchBox.scss
@@ -1,15 +1,0 @@
-.SearchBoxStyle {
-    position: absolute;
-    left: 40%;
-    padding-top: 100px;
-}
-.SearchBoxStyle {
-    color: white;
-}
-.formStyle{
-    width: 350px;
-    height: 45px;
-    border-radius: 10px;
-    padding-left: 10px;
-    border:none;
-}
